refactor: error-inteceptor move to platform library
