+++ payment/internal/model/errors.go
@@ -3,7 +3,7 @@ package model
 import (
 	"errors"
 
-	sharedErr "github.com/crafty-ezhik/rocket-factory/shared/pkg/errors"
+	sharedErr "github.com/crafty-ezhik/rocket-factory/platform/pkg/grpc/errors"
 )
 
 var (
